refactor: server structure and improve CLI handling
Moved server creation logic to a reusable function and separated CLI logic into a new `cli.py` module for better modularity and maintainability. Updated `pyproject.toml` script entry to reflect the new CLI location.
