+++ src/mcp_florence2/__main__.py
@@ -5,7 +5,6 @@
 #  This software is released under the MIT License.
 #
 #  http://opensource.org/licenses/mit-license.php
-
-from . import main
+from mcp_florence2.cli import main
 
 main()
